Serial Rx/Tx works, but not DALI testable yet due to ruined output drivers
We had VPP connected to RB4/PGM for quite some time, which ruined the
output driver of RB4 completely, and apparently also ruined the
RA4 output somewhat.

Source and Hardware is now changed to use RB3 instead of RB4.

But as RA4 seems to be damaged as well (this is the current regulator
output), we'll need a new PIC anyway.